requirements.txt
streamlit
anthropic
python-dotenv
pypdf2
python-docx
pdfplumber
pandas
openpyxl
xlrd
requests